+++ src/main/java/gift/repository/WishRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface WishRepository extends JpaRepository<Wish, Integer> {
+public interface WishRepository extends JpaRepository<Wish, Long> {
     List<Wish> findByMemberId(Long memberId);
 }
